Add RespondToApplicationHelper
