+++ src/main/controllers/RespondToApplicationController.ts
@@ -3,8 +3,7 @@ import { Response } from 'express';
 import { Form } from '../components/form/form';
 import { isFieldFilledIn } from '../components/form/validator';
 import { AppRequest } from '../definitions/appRequest';
-import { YesOrNo } from '../definitions/case';
-import { ErrorPages, PageUrls, Rule92Types, TranslationKeys } from '../definitions/constants';
+import { PageUrls, Rule92Types, TranslationKeys } from '../definitions/constants';
 import { FormContent, FormFields } from '../definitions/form';
 import { SupportingMaterialYesNoRadioValues } from '../definitions/radios';
 import { AnyRecord } from '../definitions/util-types';
@@ -13,16 +12,11 @@ import { getLogger } from '../logger';
 import { getFlagValue } from '../modules/featureFlag/launchDarkly';
 
 import { getTseApplicationDetails } from './helpers/ApplicationDetailsHelper';
-import { setUserCase } from './helpers/CaseHelpers';
-import {
-  createDownloadLink,
-  findSelectedGenericTseApplication,
-  populateDocumentMetadata,
-} from './helpers/DocumentHelpers';
-import { getResponseErrors as getApplicationResponseError } from './helpers/ErrorHelpers';
+import { createDownloadLink, populateDocumentMetadata } from './helpers/DocumentHelpers';
 import { getPageContent } from './helpers/FormHelpers';
 import { getApplicationRespondByDate } from './helpers/PageContentHelpers';
-import { getLanguageParam, returnSafeRedirectUrl } from './helpers/RouterHelpers';
+import { handlePost } from './helpers/RespondToApplicationHelper';
+import { getLanguageParam } from './helpers/RouterHelpers';
 
 const logger = getLogger('RespondToApplicationController');
 
@@ -63,33 +57,7 @@ export default class RespondToApplicationController {
   }
 
   public post = async (req: AppRequest, res: Response): Promise<void> => {
-    setUserCase(req, this.form);
-    const languageParam = getLanguageParam(req.url);
-    const selectedApplication = findSelectedGenericTseApplication(
-      req.session.userCase.genericTseApplicationCollection,
-      req.params.appId
-    );
-    if (selectedApplication === undefined) {
-      logger.error('Selected application not found');
-      return res.redirect(ErrorPages.NOT_FOUND + languageParam);
-    }
-
-    const formData = this.form.getParsedBody(req.body, this.form.getFormFields());
-    const error = getApplicationResponseError(formData);
-
-    if (error) {
-      req.session.errors = [];
-      req.session.errors.push(error);
-      const redirectUrl =
-        PageUrls.RESPOND_TO_APPLICATION_SELECTED.replace(':appId', selectedApplication.id) + languageParam;
-      return res.redirect(returnSafeRedirectUrl(req, redirectUrl, logger));
-    }
-    req.session.errors = [];
-    const redirectUrl =
-      req.session.userCase.hasSupportingMaterial === YesOrNo.YES
-        ? PageUrls.RESPONDENT_SUPPORTING_MATERIAL.replace(':appId', selectedApplication.id) + languageParam
-        : PageUrls.COPY_TO_OTHER_PARTY + languageParam;
-    return res.redirect(returnSafeRedirectUrl(req, redirectUrl, logger));
+    await handlePost(req, res, this.form, PageUrls.RESPOND_TO_APPLICATION_SELECTED, logger);
   };
 
   public get = async (req: AppRequest, res: Response): Promise<void> => {
